# contracts/api-contract.yaml
...
openapi: 3.0.3
info:
  title: kart-admin-service API
  version: 1.0.0
  description: >
    Back-office operations and fine-grained, category-scoped RBAC for the
    platform's four fixed operational categories (catalog management, coupon
    issuance, user suspension, inventory replenishment) plus a fifth
    meta-category (permission-management) that governs the other four
    (BRD §2.1 item 20, §24.1; ADR-0010). Admin Service is an orchestration/
    control-plane layer only — it never stores a second copy of Product,
    Category, Coupon, user-identity, or Inventory data (Domain Invariant #3);
    every mutation here is proxied synchronously to the owning service's own
    write API after Admin's own fine-grained grant check passes.
servers:
  - url: /v1
paths:
...
  /admin/permission-grants:
    get:
      operationId: listPermissionGrants
      summary: List permission grants, optionally filtered by principal and/or category.
      description: >
        Read path over admin_permission_grants (database-design.md). Backs
        the same "which categories can this Admin-role holder actually
        exercise" question every /admin/* handler answers internally on
        every request, exposed here for operator/audit visibility.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: principalId
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { $ref: '#/components/schemas/GrantCategory' }
        - name: includeRevoked
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Matching grant records.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/PermissionGrant' }
                  page: { type: integer }
                  pageSize: { type: integer }
                  total: { type: integer }
        '401':
          description: Not authenticated (missing/invalid Identity-issued JWT).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `permission-management` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

    post:
      operationId: issuePermissionGrant
      summary: Issue (or re-issue) a live category grant to a principal.
      description: >
        Requires the caller to hold a live `permission-management` grant
        themselves (requirement-spec.md §6 Decision item 1). Writes a new
        admin_permission_grants row and, in the same local transaction, an
        admin_actions row (action `grant.issue`), then publishes
        `AdminActionPerformed` via the Outbox pattern (edge-cases.md,
        "Admin Action Succeeds but AdminActionPerformed Is Lost"). This is a
        local-only write — no synchronous outbound call to another service
        is made for this operation.
      security:
        - bearerAuth: [admin]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [principalId, category]
              properties:
                principalId: { type: string, description: "Identity-issued principal id being granted this category (Domain Invariant #1 — no FK, Identity owns this identifier)." }
                category: { $ref: '#/components/schemas/GrantCategory' }
      responses:
        '201':
          description: Grant issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PermissionGrant' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `permission-management` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: A live (non-revoked) grant already exists for this (principalId, category) — matches the partial unique index `uq_admin_permission_grants_live` (database-design.md); revoke the existing grant before re-issuing.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/permission-grants/{grantId}/revoke:
    post:
      operationId: revokePermissionGrant
      summary: Revoke a live permission grant.
      description: >
        Requires the caller to hold a live `permission-management` grant.
        Optimistic concurrency via the grant's `version` (design-decisions.md,
        "Concurrency Control for Back-Office Writes"): the caller must supply
        the version it last read; a mismatch means another admin already
        mutated this grant. Takes effect on the affected principal's very
        next request — no token-TTL exposure window (edge-cases.md, "Stale
        Admin Permission Outliving an Identity-Side Revocation").
      security:
        - bearerAuth: [admin]
      parameters:
        - name: grantId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
        - name: If-Match
          in: header
          required: true
          description: Expected current `version` of the grant row.
          schema: { type: integer }
      responses:
        '200':
          description: Grant revoked.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PermissionGrant' }
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Caller lacks the coarse `Admin` role claim, or lacks a live `permission-management` grant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: grantId not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Stale `If-Match` version (edge-cases.md, "Concurrent Admins Editing the Same Back-Office Record") — re-read and retry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ---------------------------------------------------------------------
  # Catalog management — product half (category `catalog-management`).
  # Proxies Product Service's own write API (POST /products,
  # PUT/PATCH /products/{id} — kart-product-service/requirement-spec.md
  # §5, resolving its Open Question #4 per ADR-0010). Product's
  # deactivate route is not yet named upstream — assumed here as a
  # dedicated action route, to reconcile once Product's own contract exists.
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Identity-issued RS256 JWT carrying the `Admin` role claim
        (kart-identity-service/api-contract.yaml's TokenPair.roles), coarse-
        validated at the API Gateway before the request reaches Admin
        Service (BRD §24.1's two-tier enforcement, requirement-spec.md §2).
        Admin Service performs no separate coarse revocation-list check of
        its own (edge-cases.md, "Stale Admin Permission Outliving an
        Identity-Side Revocation") — it trusts the Gateway's per-request
        revalidation for this layer and adds only its own fine-grained,
        category-scoped grant check (admin_permission_grants) on top.
  schemas:
    GrantCategory:
      type: string
      description: >
        The six fixed categories requirement-spec.md §6 Decision item 1
        and database-design.md's CHECK constraint define — five business
        categories (`order-management` added for the "Order Management
        (Admin)" flow) plus the `permission-management` meta-category.
      enum:
        - catalog-management
        - coupon-issuance
        - user-suspension
        - inventory-replenishment
        - permission-management
        - order-management
...
    PermissionGrant:
      type: object
      description: Mirrors database-design.md's admin_permission_grants row.
      required: [grantId, principalId, category, grantedAt, grantedBy, version]
      properties:
        grantId: { type: string, format: uuid }
        principalId: { type: string }
        category: { $ref: '#/components/schemas/GrantCategory' }
        grantedAt: { type: string, format: date-time }
        grantedBy: { type: string, description: "Principal id of the granting admin, or 'seed-script' for the one-time bootstrap grant." }
        revokedAt: { type: string, format: date-time, nullable: true }
        revokedBy: { type: string, nullable: true }
        version: { type: integer, description: "Optimistic-concurrency row version (design-decisions.md)." }
...
    Problem:
      type: object
      description: RFC 7807-style problem details (matches kart-identity-service/api-contract.yaml's Problem convention).
      required: [code, message]
      properties:
        code:
          type: string
          example: permission_denied
        message:
          type: string
        details:
          type: object
          additionalProperties: true